testing code
i've begun testing but have run into an issue i don't know the answer to.
an error is being given that the dispatcher thread is not the owner of a semaphore, and cannot use it.
beside that there are 2 issues i've noted.
-tasks start at 16 instead of 0 consistently.
-the arguments "-C 1 -S 2" give the result of
"invalid fifth argument"
when it should be
"No time quantum entered"